deleted app folder, added tests

+++ contracts/EthPriceOracle.sol
@@ -7,10 +7,17 @@ contract EthPriceOracle {
 	AggregatorV3Interface internal priceFeed;
 
 	constructor() public {
+		/* the constuructor initializes the address of the AggregatorV3Interface
+		 * with the contract address of the Chainlink PriceFeed oracle
+		 */
 		priceFeed = AggregatorV3Interface(0x9326BFA02ADD2366b30bacB125260Af641031331);
 	}
 		
-	function getLatestPrice() public view returns (int) {
+	function getLatestPrice() public view returns (uint) {
+		/* the "latestRoundData" call is returning multiple values, so
+		 * we have to store them in a tuple otherwise we would only get
+		 * the first returned element
+		 */
 		(
 			uint80 roundID, 
 			int price,
@@ -18,6 +25,8 @@ contract EthPriceOracle {
 			uint timeStamp,
 			uint80 answeredInRound
 		) = priceFeed.latestRoundData();
-		return price;
+
+		// we only need the current price info so we only return the price
+		return uint(price);
 	}
 }
